Refactor memory management and improve WebAssembly interpreter.
Replaced `IWebAssemblyInterpreterMemoryArea` with `IWebAssemblyMemoryArea` for a unified memory interface, introducing `WebAssemblyHeapMemoryArea`. Enhanced `WebAssemblyInterpreterMethod` and runtime setup to support imported and local memory areas. Updated tests to validate new memory handling functionality.

+++ WebAssemblySharp/Readers/Binary/WasmBinaryReader.cs
@@ -420,7 +420,7 @@ public class WasmBinaryReader
             {
                 // Only Min
                 m_CurrentMemory.Min = -1;
-                m_CurrentMemory.Max = Int64.MaxValue;
+                m_CurrentMemory.Max = Int32.MaxValue;
             }
             else
             {
@@ -569,7 +569,7 @@ public class WasmBinaryReader
 
                 if (l_HasMax[0] == 0)
                 {
-                    l_ImportMemory.Max = Int64.MaxValue;
+                    l_ImportMemory.Max = Int32.MaxValue;
                 }
                 else
                 {
